separate states into sub-folder files
